Move SealWrapper to its own file.
Remove unused method SealWrapper.keyId().
